search component, refactoring
